Add comments to the buildin factories definitions

+++ factories/posts.php
@@ -1,15 +1,27 @@
 <?php
 
+/**
+ * Make factory `posts` definition for seeding
+ * entries of the `post` posttype.
+ */
 add_action('wp_seeder/define/factory/post', function (WPSeeder\Factory $factory) {
     $factory->define('posts', function (Faker\Generator $faker) {
         return ['post_type' => 'post'];
     });
 });
 
+/**
+ * Generate specified $count number
+ * seeds of `posts` definition.
+ */
 add_action('wp_seeder/generate/seeds', function (WPSeeder\Factory $factory) use ($count) {
     $factory->create('posts', $count);
 });
 
+/**
+ * Before running seeding, we will delete all
+ * current entries of `post` post type.
+ */
 add_action('wp_seeder/before/run', function () {
     $posts = get_posts([
         'post_type' => 'post',
@@ -21,15 +33,23 @@ add_action('wp_seeder/before/run', function () {
     }
 });
 
+/**
+ * We also have to delete attachments of every
+ * `post` entry which will be deleted.
+ */
 add_action('before_delete_post', function ($id) {
-    $media = get_children([
-        'post_parent' => $id,
-        'post_type'   => 'attachment'
-    ]);
+    global $post_type;
 
-    if(! empty($media)) {
-        foreach($media as $file) {
-            wp_delete_attachment($file->ID);
+    if ($post_type === 'post') {
+        $media = get_children([
+            'post_parent' => $id,
+            'post_type'   => 'attachment'
+        ]);
+
+        if(! empty($media)) {
+            foreach($media as $file) {
+                wp_delete_attachment($file->ID);
+            }
         }
     }
 });
